+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Added new parameter `DisableTtl` to `Restore-AzCosmosDBAccount`.
 
 ## Version 1.14.4
 * Fixed the issue that Azure.Core.AccessToken is used before assigned.